+++ src/reducer/project.js
@@ -17,9 +17,17 @@ export default handleActions(
     },
 
 
-    [ actions.resetProject ]: ( state ) => update( state, {
-      $set: DEFAULT_STATE.project
-    }),
+    [ actions.resetProject ]: ( state, { payload } ) => {
+      if ( !payload ) {
+        return update( state, {
+          $set: DEFAULT_STATE.project
+        });
+      }
+
+      return update( DEFAULT_STATE.project, {
+        $merge: payload
+      });
+    },
 
 
     [ actions.addEnv ]: ( state, { payload }) => {